web-app/requirements.txt
flask==3.1.0
pymongo==4.12.0
flask-pymongo==3.0.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-flask==1.3.0
typing_extensions==4.13.2
beautifulsoup4==4.13.4
coverage==7.8.0
pytest-cov==6.1.1